Fix invalid oxend rpc call
This call is wrong (and has always been wrong): the field should be set
in the top-level parameters, not inside `fields`.

In current dev lokinet, however, the invalid field becomes an error,
because dev tries to parse all `fields` keys with `bool` values, and
this ends up returning an error to Lokinet instead of a response, which
breaks Lokinet's service node updating code.